Scriptly pushed

+++ compile.sh
@@ -1,5 +1,13 @@
 #!/bin/sh
 
+
+if [ "$1" == "ssl" ]; then
+    cd ./modules/SSL/
+    cmake . -G "Unix Makefiles" && cmake --build . && make
+    mv ./libmod_ssl.so ../
+    exit
+fi
+
 rm -rf ./build
 rm ./modules/*.so
 
@@ -10,7 +18,7 @@ fi
 mkdir ./build && cd ./build &&
 conan install .. --build=missing &&
 cmake .. -G "Unix Makefiles" && cmake --build . &&
-make &&
+make  -j5 &&
 
 # we still in ./build
 cp ./lib/*.so ../modules/ &&
